Token consData processing
